mfinance-core MainContainer Builder refactoring in progress

+++ mfinance-core/src/main/java/com/infoshare/mfinance/core/models/bossa/Currency.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class Currency extends Investment {
 
-    public Currency(int Id, String name, List<Quotation> quotations) {
+    public Currency(long Id, String name, List<Quotation> quotations) {
         super(Id, name, quotations);
     }
