+++ mfinance-core/src/main/java/com/infoshare/mfinance/core/models/bossa/Fund.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class Fund extends Investment {
 
-    public Fund(int Id, String name, List<Quotation> quotations) {
+    public Fund(long Id, String name, List<Quotation> quotations) {
         super(Id, name, quotations);
     }
